yolov8_barcode_project/requirements.txt
ultralytics>=8.0.0
torch>=2.0.0
torchvision>=0.15.0
onnx>=1.15.0
onnxruntime>=1.16.0
opencv-python>=4.8.0
numpy>=1.24.0
pillow>=10.0.0
matplotlib>=3.7.0
pyyaml>=6.0
tqdm>=4.65.0
